some hacks to fix broken uart and replace it with ppm

+++ EBike_Software/DriveTasks.cpp
@@ -10,6 +10,6 @@ void DriveTasks::throttleTaskStart(void*){
 
 void DriveTasks::throttleTask(void*){
   float amps = 0.0f;
-  if(AuthTask::isAuthenticated()) amps = Controls::getInstance()->getThrottle()*Drive::MAX_CURRENT;
+  if(AuthTask::isAuthenticated()) amps = Controls::getInstance()->getThrottle()*(float)Drive::MAX_CURRENT;
   Drive::getInstance()->setAmps(amps);
 }
